Fix PKGMAN_CompileDir to work for out-of-tree builds
The BuildPackages.sh is always contained in the GAP source directory.
But the gaproot it expects as argument actually must point at the GAP
build directory, where sysinfo.gap is located. In regular GAP builds,
these two directories are the same. But the whole point of out-of-tree
builds is to allow for for a separate build dir.