do not call `Error` in `GetPackageURLs`
Most of the `Error` (actually `ErrorNoReturn`) calls in the code
of PackageManager belong to argument checks.
These are fine from the viewpoint of programmatic access to
PackageManager's functions.

The only exceptions are two places in `GetPackageURLs`,
which are reached if either `PKGMAN_DownloadURL` fails
or the downloaded file is corrupted.
These error are problematic for example when Julia code wants to use
PackageManager to install GAP packages at runtime,
see oscar-system/Oscar.jl/issues/1364.

The proposed changes replace the `ErrorNoReturn` calls
by returning the result record with a new component `success`
that is set to `false`.
In the three situations where `GetPackageURLs` is called,
this component gets evaluated.

As far as I see, the documentation need not be changed.
First of all, the function `GetPackageURLs` is undocumented.
The three functions that call it (`InstallPackageFromName`,
`PKGMAN_InstallDependencies`, `UpdatePackage`) do anyhow return
results that indicate failures, and they do not promise error messages
if one is offline.

In order to inform the user about the reason for the failures,
now info messages are printed by `GetPackageURLs`

(Would there be any way to *test* the offline behaviour of PackageManager?)