changed the dependencies
Up to now, curlInterface had been a suggested package.
Now we have to turn utils into a needed package,
when then has curlInterface as a suggested package.

Note that PackageManager had provided its own download functionality
for which curlInterface was one possible tool.
Now the decision about download tools gets outsourced to utils.

(As a consequence, PackageManager cannot be loaded if the extended
version of utils is not available --and currently no released version
of utils has the required functionality.)