fix `InstallPackage` for prescribed versions
If an exact version number (one starting with `=`) is prescribed
in `InstallPackage`, the exactness was ignored up to now.
With the change, an available version with larger version number
is not regarded as matching the version condition in this case,
and the return value is `false`.

This affects `InstallPackageByName` and `InstallPackageByInfo`.
(The possibility to prescribe a version number in the latter
function is in fact not documented.)